fix: message creation flow
