+++ libs/editor/src/lib/extensions/keyboard-handler.ts
@@ -1,0 +1,20 @@
+import { Extension } from '@tiptap/core';
+
+export const KeyboardHandler = Extension.create({
+  name: 'keyboardHandler',
+  addKeyboardShortcuts() {
+    return {
+      'Shift-Enter': ({ editor }) => {
+        return editor.commands.first(({ commands }) => [
+          () => commands.newlineInCode(),
+          // () => commands.splitListItem('listItem'),
+          () => commands.createParagraphNear(),
+          () => commands.liftEmptyBlock(),
+          () => commands.splitBlock(),
+        ]);
+      },
+      Enter: () => true,
+      'Mod-Enter': () => true,
+    };
+  },
+});
